do not modify existing settings if not specified in config string.
Add one more unit test,

ArrowSetup errorCheck should exempt arguments if
config.error1006.exempt is specified but keep other existing exemption
settings if it is not specified in config string

For example, by default capabilities is exempted.
When user only provide a config string which does not modify the
capabilities, the existing exemption settings of capabilities should be
preserved.